# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeAI - 코딩 AI 어시스턴트 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface User {
  id: number;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface MessageResponse {
  message: string;
}

export interface Project {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  messageCount?: number;
  fileCount?: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface GeneratedFile {
  id: number;
  messageId: number;
  filename: string;
  fileType: string;
  sizeBytes: number;
  createdAt: string;
}

export interface Message {
  id: number;
  projectId: number;
  role: MessageRole;
  content: string;
  files?: GeneratedFile[];
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface ProjectStats {
  totalProjects: number;
  totalMessages: number;
  totalFiles: number;
  recentProjects?: Project[];
}