# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeAI - 코딩 AI 어시스턴트 API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  email: string;
  createdAt: Date;
}